+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [#34] Fixed a permission issue where the remote registry trys to write to a non-privileged cache dir.
 
 ## [v0.9.0] - 2022-07-13
 ### Added